Organized test data for processing multisearch results
